+++ app/src/scenes/Persons/Person.js
@@ -184,7 +184,7 @@ const Person = ({ route, navigation }) => {
       const keepGoing = await new Promise((res) => {
         Alert.alert(
           'Voulez-vous continuer la suppression ?',
-          'Des données médicales sont associées à cette personne. Si vous la supprimez, ces données seront également effacées. Vous n’avez pas accès à ces données médicales car vous n’êtes pas un professionnel de santé. Voulez-vous supprimer cette personne et toutes ses données ?',
+          'Des données médicales sont associées à cette personne. Si vous la supprimez, ces données seront également effacées. Vous n’avez pas accès à ces données médicales car vous n’êtes pas un·e professionnel·le de santé. Voulez-vous supprimer cette personne et toutes ses données ?',
           [
             { text: 'Annuler', style: 'cancel', onPress: () => res(false) },
             { text: 'Continuer', style: 'destructive', onPress: () => res(true) },