+++ dashboard/src/scenes/user/view.js
@@ -114,10 +114,10 @@ const View = () => {
                       handleChange({ target: { value: !values.healthcareProfessional, name: 'healthcareProfessional' } });
                     }}
                   />
-                  Professionnel de santé
+                  Professionnel·le de santé
                 </Label>
                 <div>
-                  <small className="text-muted">Un professionnel de santé a accès au dossier médical complet des personnes.</small>
+                  <small className="text-muted">Un·e professionnel·le de santé a accès au dossier médical complet des personnes.</small>
                 </div>
               </Col>
             </Row>